rename : loading 컴포넌트 common폴더로 이동
